fix(@desktop/chat): make scroll to message from search work again
We were trying to call `scrollToMessage` of `undefined`. Seems like the `Loader`
used in the chat messages view was recently introduced, causing the property lookup
in the API call to be incomplete.

This commit fixes this by exposing the loader on the chat messages column and then accessing
the `scrollToMessage` API from there.

Closes: #3362