fix(@desktop/browser): fix paste button in favorites modal
This fixes the base button in the favorites modal by replacing the legacy
`Input` component with StatusQ's `StatusInput` and `StatusButton` components.

It also updates the validation mechanism to use the one provided by StatusQ.
The `height` of the modal has been adjusted due to its content growing when
validation errors are shown.

This will go away as soon as we replace `ModalPopup` with `StatusModal`.

Fixes #3977