fix: ensure last active community is reset when leaving community
There's a bug where the last active community will be set in local settings
and not removed when a user leaves a community.

This causes the app to start on an invalid screen/state as it still tries to
render the community view for a community that the user has just left.

This commit ensure we're resetting the last active community in the local settings
when a users leaves the community.

Fixes: #4094