fix: ensure app sections work correctly
We've recently removed timeline related code from the application.
It seems that either a) new changes have been introduced related to timeline
despite it being removed from the application, or b) a few more changes
where left out as part of the new architecture.

This causes the app to misbehave when selecting sections in the application
nav bar.

This commit fixes this by removing leftover timeline related code.