refactor(@desktop/general): old code disconnected
The old code placed in `src/app` is still not deleted, due to easier refactoring
the rest of the app, but non of the part of the app refers to it. Also corresponding
directories are renamed, to spot them easier, and at the end of the refactor we
will just simply delete them in a single commit.

Renamed:
- `src/chat` to `src/REMOVE-chat`
- `src/keycard` to `src/REMOVE-keycard`
- `src/node` to `src/REMOVE-node`
- `src/profile` to `src/REMOVE-profile`
- `src/utilsView` to `src/REMOVE-utilsView`
- `src/wallet` to `src/REMOVE-wallet`