fix: make changing local nicknames of contacts work again
This fixes #4173 again. The problem is that changing the local nickname
of a contact now uses new APIs in Status Desktop which emit new events that
older code isn't aware off. Namely, the event handling in the chat section
doesn't get informed anymore that the contact has changed, so it won't update
the view accordingly.

This commit fixes it by subscribing to that newly introduced event
and then updating the view with the payload data.

Notice that this is a quick fix and will most likely be obsolete, once
the chat section has moved to the new architecture as well.

Fixes #4173