fix(chat/mentions) fixes in mentions behavior
Changed cursor behavior to jump over text when it
reaches a mention, either on left or right side.
Also made mention to behave as link when hovered
with mouse so that it cannot be clicked and place
cursor inside it.

Also fixed keyboard focus to navigate only in the
suggestions list when this is opened.

Closes #4423