fix(Chat/Stickers): Recent stickers icon is missing
It has been changed footer layout from Item to row.
The colour of the Add button has been changed to match the rest of the style.
The recent tab control has been changed to be consistent with other popups like gifts or emojis and the used component directly controls the different themes.

Fixes #5458