fix(CommunityService): ensure communitySettings aren't mistakenly overwritten
When communities are edited, status-go sends a signal with the community
description, to ensure the network is updated to the latest changes.

However, that community description doesn't include `communitySettings`.
When we then compose the `CommunityDto` object from that signal, it will have
a default certain settings to `false` which might actually be `true`.

The result is that data in memory is incorrect.

To ensure we don't run into this issue, we check whether the attached
settings have an id, if they don't, we know the settings do not
originate from the signal data, meaning we can use the settings we
already have.