fix(sync): handle syncronization of user status
This fixes a bug where the new user status set by the logged in user
would not be broadcast because the corresponding setting would prevent
status-go from doing so.

The fix is done by always enabling broadcasting the user status, then
disabling it again right after setting, in case it was set to "offline".

This commit also ensures that, when current user status updates are
broadcast, the updates are handled across paired devices in real-time.

Partially addresses #5201