fix(Privacy): handle sync setting of `show-profile-pictures-to`
There are a bunch of settings which are synced an not handled by
desktop, one of them being the `profile-pictures-show-to` setting.

This commit introduces a new `SettingsFieldDto` so we can react to
message signals that include individual setting fields that originate
from syncing.

The first setting hanlded is the one mentioned above, so a new
application signal is introduced to inform the app that there was an
update in that particular setting, so it can re-render the view.

Partially addresses #5201