fix(Profile): ensure backup seedphrase warning disappears
The warning to back up the seedphrase should disappear after either
dismissing the message, or in fact backing up the seed phrase.

The latter was not happening because the banner didn't take that
information into account. This commit fixes this by ensuring we let the
banner know whether the seed phrase was already backed up.

Notice that we don't want to render the banner when the seed phrase has
*not* been backed up yet, but was dismissed by the user.

Fixes #6255