fix(@desktop/chat): sort users by online status
Users model has to be sorted by online status first, otherwise
`UserListPanel` sections will be duplicated. That's because:

> Note: Adding sections to a ListView does not automatically re-order
the list items by the section criteria. If the model is not ordered by
section, then it is possible that the sections created will not be
unique; each boundary between differing sections will result in a
section header being created even if that section exists elsewhere.

fixes: #6563