fix(Language): App crashes when language is changed
Enabled restart info popup after language is changed --> Workaround to temporary resolve the crash we have when language is changed (`startupModule` is null and some qml bindings are still calling this dead pointer) so, change language will not retranslate and instead, also on mac and win, the app will need a restart to apply the new language.

Disabled retranslations for all os on the corresponding service.

Will temporary fix #7709