refactor(communities): allow removing items from import file list
There was a requested design change where no longer wanted to have
checkboxes to decide which files will be included for a discord import,
but rather have an "X" button that enables users to remove items.

This commit implements this refactor.

In addition, it ensures that the already loaded discord categories and
channels that have been extracted from validation, are kept in sync as
well.

Meaning, if a user removes a file from the file list, the corresponding
channel will be removed as well.

If there's not channel in a given category, the category will be removed
as well.

Closes #8125 #8126