fix(StatusQ) scrolling on macOS if mouse pointer is over StatusBaseInput
This commit fixes a scrolling on macOS using trackpad in the context
of Add Wallet Account using seed-phrase.

Considerations:
The scrolling of advanced input with seed-phrase selected was failing
when mouse pointer was moving over a word phrase.

Scrolling of the popup content in AddAccountModal would not work
when the mouse pointer was over a StatusSeedPhraseInput form
the ImportSeedPhrasePanel

The identified root cause is StatusBaseInput's flicker which captures
the mouse wheel event. It seems the trackpad wheel events are handled
differently on macOS by the flicker and not propagated to the parent.

Fixes #5355