fix: prevent app from crash due to missing type variants
Alright this was very painful to figure out.

Because desktop didn't know about the notification type for
`ContactVerification`, the signal processing for incoming contact
verifications would fail.

At the same time, the new message content type was missing a type
variant for content verification as well, which would cause the app to
crash at runtime for both, sender and receiver of verification requests.

Fixes #8205