tests(suite_messaging): properly check if reply message has reply
Prior to this commit we were merely checking if the message in question
was the last one.

With this commit we now check whether there's a reply in the first place
and whether the reply-to-message is the correct one.

We also distinguish whether the message we're replying to is the one
from the loggedin user (and not someone else's)