fix(chat/messages): prevent multiple same cursor fetches
It may happen `asyncLoadMoreMessagesForChat` is called many times in a
row. Previously it did not check whether there is any pending fetch for
a given chat, which resulted in many same cursors fetches from status-go.
Furthermore, in this case, `lastUsedMsgCursor` ended up with the same
value as `msgCursor` resulting in wrong state where there are no more
messages to fetch.

- Simplified message loading code by introducing MessageCursor helper
  type
- Ensured messages are not loaded twice for the same cursor

fixes: #8602