fix(chat/messages): set clock value for new messages marker
New messages marker had a clock value of "0" before. Since all
messages are inserted based on the clock value, new messages marker
would cause other "0"-valued clock items to be inserted after it,
effectively making chat header being displayed in the middle of the
chat.

Setting new messages marker clock value to the clock of the message it
points to solves the issue.

fixes: #8955