fix(@desktop/general): crash during profile keypair migration
Happens occasionally that the app crashes during the profile keypair migration flow.
It happens cause during that flow in one moment the app gets locked, cause encryption
gets changed, but some parts of the app may still try to get some data from the db (as an
action for some async event or so). That action results as the app crash. In such cases
an exception will be thrown, but it needs to be handled in functions which are doing a call.

`fetchAccounts` is guarded in this commit, cause crash was initiated there, cause it didn't
handle a call appropriately using try/catch block.