fix(StatusMessage): don't render message text of image message
Image messages have a text attached to them that is typically rendered
in app notifications (as the actual image won't fit in there).

That text was accidentally rendered as part of the reply.

This commit adds a check such that we don't render the message text
for messages of contact type image.

Closes #9383