fix(StatusChatList): Enable StatusChatList sizing and disable scroll filtering when the listview.height == listview.contentHeight
Fixing #9460 Community -> Channels list: can't scroll list of channels with touchpad
The root-cause of this issue is that we use a Listview with height == contentHeight and delegate the scrolling to a ScrollView. The reasoning is that we can attach other items to the bottom of the listview and delegating the scroll to a ScrollView will enable us to scroll the listview+other items attached at the bottom at the same time.
Fixing it by setting ListView.interactive to false when the listview cannot be scrolled.
Also fixed the listview height so that it can also be used as a scrollable listview.