remove task-runner dependency
As of #9596, task-runner is no longer used.

Due to how the build system works, task-runner was providing nim-stew,
nim-bearssl and nim-chronos to the project, even though these are also
top-level dependencies - effectively this means there were multiple
versions of these libraries in the source tree and it is not entirely
intuitive which is actually used - adding to the confusion, the
submodule was pinning different versions.

This PR removes task-runner and resets the other dependencies to the
versions in task-runner thus minimising collateral changes.

A duplicate copy of bearssl is also removed.