fix(permissions): ensure permissions model is rebuilt when permission was deleted
This is needed because there's cases where the deletion of a permission
might cause the current user to no longer be eligible to join
a community.

Community has the follwoing permissions:
- own 2 ETH and 1 FOO_NFT or
- own 1 FOO_NFT

User owns: 1 FOO_NFT

^ The above would make the user eligible to request access to join.

Now if the community removes the second permission, the user now no
longer fulfills the requirements so we need to ensure the UI
permits the user to request access.

Rebuilding the permissions model (including its token criteria) sets the
necessary flags in the UI automatically.